requirements.txt
python-telegram-bot==13.8.1
mysql-connector-python==8.0.27